src/requirements.txt
aioredis==1.2.0
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
backcall==0.1.0
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
decorator==4.4.0
Django==2.2.2
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
djangorestframework==3.9.4
djangorestframework-bulk==0.2.1
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
msgpack==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Twisted==19.2.1
txaio==18.8.1
wcwidth==0.1.7
zope.interface==4.6.0
